+++ src/main/java/java/nio/DirectLongBufferU.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2000, 2011, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2000, 2013, Oracle and/or its affiliates. All rights reserved.
  * ORACLE PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
  *
  *
@@ -242,7 +242,7 @@ class DirectLongBufferU
     }
 
     private long ix(int i) {
-        return address + (i << 3);
+        return address + ((long)i << 3);
     }
 
     public long get() {
@@ -253,9 +253,15 @@ class DirectLongBufferU
         return ((unsafe.getLong(ix(checkIndex(i)))));
     }
 
+
+
+
+
+
+
     public LongBuffer get(long[] dst, int offset, int length) {
 
-        if ((length << 3) > Bits.JNI_COPY_TO_ARRAY_THRESHOLD) {
+        if (((long)length << 3) > Bits.JNI_COPY_TO_ARRAY_THRESHOLD) {
             checkBounds(offset, length, dst.length);
             int pos = position();
             int lim = limit();
@@ -267,13 +273,13 @@ class DirectLongBufferU
 
             if (order() != ByteOrder.nativeOrder())
                 Bits.copyToLongArray(ix(pos), dst,
-                                          offset << 3,
-                                          length << 3);
+                                          (long)offset << 3,
+                                          (long)length << 3);
             else
 
                 Bits.copyToArray(ix(pos), dst, arrayBaseOffset,
-                                 offset << 3,
-                                 length << 3);
+                                 (long)offset << 3,
+                                 (long)length << 3);
             position(pos + length);
         } else {
             super.get(dst, offset, length);
@@ -323,7 +329,7 @@ class DirectLongBufferU
 
             if (srem > rem)
                 throw new BufferOverflowException();
-            unsafe.copyMemory(sb.ix(spos), ix(pos), srem << 3);
+            unsafe.copyMemory(sb.ix(spos), ix(pos), (long)srem << 3);
             sb.position(spos + srem);
             position(pos + srem);
         } else if (src.hb != null) {
@@ -347,7 +353,7 @@ class DirectLongBufferU
 
     public LongBuffer put(long[] src, int offset, int length) {
 
-        if ((length << 3) > Bits.JNI_COPY_FROM_ARRAY_THRESHOLD) {
+        if (((long)length << 3) > Bits.JNI_COPY_FROM_ARRAY_THRESHOLD) {
             checkBounds(offset, length, src.length);
             int pos = position();
             int lim = limit();
@@ -358,12 +364,16 @@ class DirectLongBufferU
 
 
             if (order() != ByteOrder.nativeOrder())
-                Bits.copyFromLongArray(src, offset << 3,
-                                            ix(pos), length << 3);
+                Bits.copyFromLongArray(src,
+                                            (long)offset << 3,
+                                            ix(pos),
+                                            (long)length << 3);
             else
 
-                Bits.copyFromArray(src, arrayBaseOffset, offset << 3,
-                                   ix(pos), length << 3);
+                Bits.copyFromArray(src, arrayBaseOffset,
+                                   (long)offset << 3,
+                                   ix(pos),
+                                   (long)length << 3);
             position(pos + length);
         } else {
             super.put(src, offset, length);
@@ -381,7 +391,7 @@ class DirectLongBufferU
         assert (pos <= lim);
         int rem = (pos <= lim ? lim - pos : 0);
 
-        unsafe.copyMemory(ix(pos), ix(0), rem << 3);
+        unsafe.copyMemory(ix(pos), ix(0), (long)rem << 3);
         position(rem);
         limit(capacity());
         discardMark();
